requirements.txt
streamlit>=1.25.0
PyMuPDF>=1.24.0
Pillow>=10.0.0
